Make meo tests work

+++ server/src/main/java/io/crate/planner/optimizer/iterative/Memo.java
@@ -19,7 +19,7 @@
  * software solely pursuant to the terms of the relevant commercial agreement.
  */
 
-package io.crate.planner.iterative;
+package io.crate.planner.optimizer.iterative;
 
 
 import java.util.ArrayList;
@@ -34,6 +34,7 @@ import static java.util.Objects.requireNonNull;
 
 import io.crate.planner.operators.LogicalPlan;
 import io.crate.planner.operators.LogicalPlanIdAllocator;
+import io.crate.planner.operators.LogicalPlanVisitor;
 
 /**
  * Stores a plan in a form that's efficient to mutate locally (i.e. without
@@ -61,8 +62,7 @@ import io.crate.planner.operators.LogicalPlanIdAllocator;
  * Groups are reference-counted, and groups that become unreachable from the root
  * due to mutations in a subtree get garbage-collected.
  */
-public class Memo
-{
+public class Memo {
     private static final int ROOT_GROUP_REF = 0;
 
     private final LogicalPlanIdAllocator idAllocator;
@@ -72,52 +72,47 @@ public class Memo
 
     private int nextGroupId = ROOT_GROUP_REF + 1;
 
-    public Memo(LogicalPlanIdAllocator idAllocator, LogicalPlan plan)
-    {
+    public Memo(LogicalPlanIdAllocator idAllocator, LogicalPlan plan) {
         this.idAllocator = idAllocator;
         rootGroup = insertRecursive(plan);
         groups.get(rootGroup).incomingReferences.add(ROOT_GROUP_REF);
     }
 
-    public int getRootGroup()
-    {
+    public int getRootGroup() {
         return rootGroup;
     }
 
-    private Group getGroup(int group)
-    {
-        return groups.get(group);
-    }
-
-    public LogicalPlan getNode(int group)
-    {
+    public LogicalPlan getNode(int group) {
         return getGroup(group).membership;
     }
 
-    public LogicalPlan resolve(GroupReference groupReference)
-    {
+    public LogicalPlan resolve(GroupReference groupReference) {
         return getNode(groupReference.groupId());
     }
 
-    public LogicalPlan extract()
-    {
+    public LogicalPlan extract() {
         return extract(getNode(rootGroup));
     }
 
-    private LogicalPlan extract(LogicalPlan node)
-    {
+    private LogicalPlan extract(LogicalPlan node) {
         return resolveGroupReferences(node, Lookup.from(planNode -> Stream.of(this.resolve(planNode))));
     }
 
-    public LogicalPlan replace(int groupId, LogicalPlan node, String reason)
-    {
+    private Group getGroup(int group) {
+        if (!groups.containsKey(group)) {
+            throw new IllegalStateException();
+        }
+        return groups.get(group);
+    }
+
+
+    public LogicalPlan replace(int groupId, LogicalPlan node, String reason) {
         Group group = getGroup(groupId);
         LogicalPlan old = group.membership;
 
         if (node instanceof GroupReference) {
             node = getNode(((GroupReference) node).groupId());
-        }
-        else {
+        } else {
             node = insertChildrenAndRewrite(node);
         }
 
@@ -128,9 +123,7 @@ public class Memo
     }
 
 
-
-    private void incrementReferenceCounts(LogicalPlan fromNode, int fromGroup)
-    {
+    private void incrementReferenceCounts(LogicalPlan fromNode, int fromGroup) {
         Set<Integer> references = getAllReferences(fromNode);
 
         for (int group : references) {
@@ -138,34 +131,34 @@ public class Memo
         }
     }
 
-    private void decrementReferenceCounts(LogicalPlan fromNode, int fromGroup)
-    {
+    private void decrementReferenceCounts(LogicalPlan fromNode, Integer fromGroup) {
         Set<Integer> references = getAllReferences(fromNode);
 
         for (int group : references) {
             Group childGroup = groups.get(group);
+            if (!childGroup.incomingReferences.remove(fromGroup)) {
+                throw new IllegalStateException("Reference to remove not found");
+            }
+            ;
             if (childGroup.incomingReferences.isEmpty()) {
                 deleteGroup(group);
             }
         }
     }
 
-    private Set<Integer> getAllReferences(LogicalPlan node)
-    {
+    private Set<Integer> getAllReferences(LogicalPlan node) {
         return node.sources().stream()
             .map(GroupReference.class::cast)
             .map(GroupReference::groupId)
             .collect(Collectors.toSet());
     }
 
-    private void deleteGroup(int group)
-    {
+    private void deleteGroup(int group) {
         LogicalPlan deletedNode = groups.remove(group).membership;
         decrementReferenceCounts(deletedNode, group);
     }
 
-    private LogicalPlan insertChildrenAndRewrite(LogicalPlan node)
-    {
+    private LogicalPlan insertChildrenAndRewrite(LogicalPlan node) {
         return node.replaceSources(
             node.sources().stream()
                 .map(child -> new GroupReference(
@@ -175,8 +168,7 @@ public class Memo
                 .collect(Collectors.toList()));
     }
 
-    private int insertRecursive(LogicalPlan node)
-    {
+    private int insertRecursive(LogicalPlan node) {
         if (node instanceof GroupReference) {
             return ((GroupReference) node).groupId();
         }
@@ -190,29 +182,51 @@ public class Memo
         return group;
     }
 
-    private int nextGroupId()
-    {
+    private int nextGroupId() {
         return nextGroupId++;
     }
 
-    public int getGroupCount()
-    {
+    public int getGroupCount() {
         return groups.size();
     }
 
-    private static final class Group
-    {
-        static Group withMember(LogicalPlan member)
-        {
+    private static final class Group {
+        static Group withMember(LogicalPlan member) {
             return new Group(member);
         }
 
         private LogicalPlan membership;
         private final List<Integer> incomingReferences = new ArrayList<>();
 
-        private Group(LogicalPlan member)
-        {
+        private Group(LogicalPlan member) {
             this.membership = requireNonNull(member, "member is null");
+        }
+    }
+
+    public static LogicalPlan resolveGroupReferences(LogicalPlan node, Lookup lookup) {
+        requireNonNull(node, "node is null");
+        return node.accept(new ResolvingVisitor(lookup), null);
+    }
+
+    private static class ResolvingVisitor extends LogicalPlanVisitor<Void, LogicalPlan> {
+        private final Lookup lookup;
+
+        public ResolvingVisitor(Lookup lookup) {
+            this.lookup = requireNonNull(lookup, "lookup is null");
+        }
+
+        @Override
+        public LogicalPlan visitPlan(LogicalPlan node, Void context) {
+            List<LogicalPlan> children = node.sources().stream()
+                .map(child -> child.accept(this, context))
+                .collect(Collectors.toList());
+
+            return node.replaceSources(children);
+        }
+
+        @Override
+        public LogicalPlan visitGroupReference(GroupReference node, Void context) {
+            return lookup.resolveGroup(node).findFirst().get().accept(this, context);
         }
     }
 }
